requirements.txt
crewai==0.80.0
pandas==2.2.3
pysqlite3_binary
streamlit==1.40.1
openpyxl